Add extra-file extra-cmd extra-package options
When creating an image the `--extra-package` option can be used to
install an additional package into the filesystem.

Similarly the `--extra-file` option can be used to install
an additional file from `overlay/` into the filesystem.

Finally the `--extra-cmd` option can be used to run a command on the
filesystem.

All of these option can be specified multiple times.